# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RecruitIQ API — recruitment knowledge management system
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List candidates with optional filters
      parameters:
        - name: roleCategoryFilter
          in: query
          schema:
            type: string
        - name: statusFilter
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CandidateWithPreferences"

    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Create a new candidate with preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateInput"
      responses:
        "201":
          description: Created candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateWithPreferences"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /candidates/check-duplicate:
    post:
      operationId: checkCandidateDuplicate
      tags: [candidates]
      summary: Check if a candidate already exists by email or phone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DuplicateCheckInput"
      responses:
        "200":
          description: Duplicate check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DuplicateCheckResult"
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get candidate by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateWithPreferences"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateUpdate"
      responses:
        "200":
          description: Updated candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateWithPreferences"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteCandidate
      tags: [candidates]
      summary: Delete a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /candidates/{id}/preferences:
    get:
      operationId: getCandidatePreferences
      tags: [preferences]
      summary: Get preferences for a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidatePreference"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: upsertCandidatePreferences
      tags: [preferences]
      summary: Create or update preferences for a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidatePreferenceInput"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidatePreference"
  /candidates/{id}/job-statuses:
    get:
      operationId: getCandidateJobStatuses
      tags: [status]
      summary: Get all job statuses for a specific candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of job statuses for candidate
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CandidateJobStatusWithJob"

  # ── Job Openings ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Candidate:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        resumeText:
          type: ["string", "null"]
        skills:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
          enum: [resume, chat, voice, mixed, null]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CandidateWithPreferences:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        resumeText:
          type: ["string", "null"]
        skills:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        preferences:
          $ref: "#/components/schemas/CandidatePreference"
    CandidateInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        resumeText:
          type: string
        skills:
          type: string
        source:
          type: string
        preferences:
          $ref: "#/components/schemas/CandidatePreferenceInput"
    CandidateUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        resumeText:
          type: string
        skills:
          type: string
        source:
          type: string
        preferences:
          $ref: "#/components/schemas/CandidatePreferenceInput"
    DuplicateCheckInput:
      type: object
      properties:
        email:
          type: string
        phone:
          type: string
    DuplicateCheckResult:
      type: object
      required: [isDuplicate]
      properties:
        isDuplicate:
          type: boolean
        existingCandidate:
          $ref: "#/components/schemas/Candidate"
...
    CandidatePreference:
      type: object
      required: [id, candidateId, updatedAt]
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        desiredRole:
          type: ["string", "null"]
        desiredRoleCategory:
          type: ["string", "null"]
        salaryExpectation:
          type: ["string", "null"]
        noticePeriod:
          type: ["string", "null"]
        wfhPreference:
          type: ["string", "null"]
        relocationWillingness:
          type: ["boolean", "null"]
        shiftPreference:
          type: ["string", "null"]
        careerInterests:
          type: ["string", "null"]
        lastVerifiedAt:
          type: ["string", "null"]
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CandidatePreferenceInput:
      type: object
      properties:
        desiredRole:
          type: string
        desiredRoleCategory:
          type: string
        salaryExpectation:
          type: string
        noticePeriod:
          type: string
        wfhPreference:
          type: string
        relocationWillingness:
          type: boolean
        shiftPreference:
          type: string
        careerInterests:
          type: string
        lastVerifiedAt:
          type: string
          format: date-time

    # ── Job Openings ──────────────────────────────────────────────────────
    JobOpening:
      type: object
      required: [id, title, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        roleCategory:
          type: ["string", "null"]
        requirements:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    CandidateJobStatusWithJob:
      type: object
      required: [id, candidateId, jobId, status, updatedAt]
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        jobId:
          type: integer
        status:
          type: string
          enum: [matched, rejected, open_for_future, interested, contacted]
        updatedBy:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time
        job:
          $ref: "#/components/schemas/JobOpening"
...