# lib/api-zod/src/generated/types/candidatePreference.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RecruitIQ API — recruitment knowledge management system
 * OpenAPI spec version: 0.1.0
 */

export interface CandidatePreference {
  id: number;
  candidateId: number;
  /** @nullable */
  desiredRole?: string | null;
  /** @nullable */
  desiredRoleCategory?: string | null;
  /** @nullable */
  salaryExpectation?: string | null;
  /** @nullable */
  noticePeriod?: string | null;
  /** @nullable */
  wfhPreference?: string | null;
  /** @nullable */
  relocationWillingness?: boolean | null;
  /** @nullable */
  shiftPreference?: string | null;
  /** @nullable */
  careerInterests?: string | null;
  /** @nullable */
  lastVerifiedAt?: Date | null;
  updatedAt: Date;
}