# lib/api-zod/src/generated/types/candidateWithPreferences.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RecruitIQ API — recruitment knowledge management system
 * OpenAPI spec version: 0.1.0
 */
import type { CandidatePreference } from './candidatePreference';

export interface CandidateWithPreferences {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  resumeText?: string | null;
  /** @nullable */
  skills?: string | null;
  /** @nullable */
  source?: string | null;
  createdAt: Date;
  updatedAt: Date;
  preferences?: CandidatePreference;
}