# lib/api-zod/src/generated/types/jobOpening.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RecruitIQ API — recruitment knowledge management system
 * OpenAPI spec version: 0.1.0
 */

export interface JobOpening {
  id: number;
  title: string;
  /** @nullable */
  roleCategory?: string | null;
  /** @nullable */
  requirements?: string | null;
  createdAt: Date;
}